Refactor: foregroundService 시작 진입점 수정
